data-generator/requirements.txt
faker==30.0.0
kafka-python==2.0.2
